Add metalava 1.0.0-SNAPSHOT
Metalava has a groupId, artifactId, and a version, but it's not released to any Maven repo nor is the version particularly meaningful. As a result, the version is treated as a SNAPSHOT and the 'shadow' classifier is used as it's a fat jar (i.e., it contains its dependencies inside of it).

Change-Id: I9c750f27a5718a5aef6a4c0acfb28e8ce6707581