Created License files for robolectric dependencies
The license filed were created manually based on the license
described in the pom file of each project. The licenses were
either Apache or MIT and there were 3 exceptions:

- Creative Commons license for backport-util-concurrent.
- Copyright (c) 2000 - 2018 The Legion of the Bouncy Castle Inc
for bcprov-jdk15on.
- Codehaus classworlds license for classworlds.

Robolectric team mentioned that if we are not using robolectric
in any different way then there should be no issue with licensing
since it's widely used at Google.

Test: ./gradlew checkExternalLicenses in platform/support
Bug: 122440485
Change-Id: I3454d7b46a26a2b77cd9d8bf6ab45346b7e4f688